Inline Account Registration forms and move actions and hooks to feature folder
