+++ application/account-management/WebApp/pages/register/expired.tsx
@@ -1,7 +1,10 @@
 import { createFileRoute } from "@tanstack/react-router";
 import { HorizontalHeroLayout } from "@/shared/ui/layout/HorizontalHeroLayout";
-import { ErrorMessage } from "@/shared/ui/auth/ErrorMessage";
-import { VerificationCodeExpiredMessage } from "@/shared/ui/auth/VerificationCodeExpiredMessage";
+import { ErrorMessage } from "@/shared/ui/components/ErrorMessage";
+import Timeout from "@spectrum-icons/illustrations/Timeout";
+import { Link } from "@repo/ui/components/Link";
+import { useRegistration } from "./-shared/actions";
+import { Content, Heading, IllustratedMessage } from "@repo/ui/components/IllustratedMessage";
 
 export const Route = createFileRoute("/register/expired")({
   component: () => (
@@ -15,3 +18,18 @@ export const Route = createFileRoute("/register/expired")({
     </HorizontalHeroLayout>
   )
 });
+
+export function VerificationCodeExpiredMessage() {
+  const { accountRegistrationId } = useRegistration();
+
+  return (
+    <IllustratedMessage>
+      <Timeout />
+      <Heading>Error: Verification code expired</Heading>
+      <Content>
+        The verification code you are trying to use has expired for Account Registration ID: {accountRegistrationId}
+      </Content>
+      <Link href="/register">Try again</Link>
+    </IllustratedMessage>
+  );
+}
